Support prepending randomtemp to a call, similar to ccache
Some versions of CMake use the nvcc path to find the cuda toolkit, so
using randomtemp to spoof the compiler will not work. With this PR you
can give CMake the real `nvcc` but use `CMAKE_CUDA_COMPILER_LAUNCHER`
to prepend `randomtemp` to the command line when compiling, just as
you can do with `ccache`.